update_2_Task2.py
Added some comments and change "int" to "float" to expand the possibilities of the program.

+++ Task2.py
@@ -1,4 +1,5 @@
+#This is a simple program that allows user calculate an area of a circle.
 import math
-radius = int(input("Radius:"))
+radius = float(input("Radius:"))
 area = math.pi * math.pow(radius, 2)
 print("Area is: ", area)
